+++ packages/cdk/lambda-python/generic-agent-core-runtime/src/__init__.py
@@ -1,0 +1,1 @@
+# Empty init file for src package
